On avance

+++ PushSwap/utils/parsing.c
@@ -6,7 +6,7 @@
 /*   By: tibras <tibras@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/29 16:22:21 by tibras            #+#    #+#             */
-/*   Updated: 2025/12/04 18:40:20 by tibras           ###   ########.fr       */
+/*   Updated: 2025/12/05 21:10:46 by tibras           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int	ft_init_str(t_list **stack, char *str)
 		nb = ft_atol(arr[i]);
 		if (ft_overint(nb) || !ft_is_only((const char *)arr[i], ft_isdigit))
 			return (ft_full_free(arr, stack, ft_node_del));
-		n_new = ft_node_init(nb);
+		n_new = ft_node_init((int)nb);
 		if (!n_new)
 			return (ft_full_free(arr, stack, ft_node_del));
 		l_new = ft_lstnew(n_new);
